Consolidated API to enforce usage pattern.
The API should be called in the following order (example shown):

   1. AsyncStackManager::AddSerial(portName)
   2. AsyncStackManager::AddMaster(portName, stackName)
   3. AsyncStackManager::AddVtoChannel(stackName, channelId)
   4. StartVtoRouter(stackName)

The VtoRouter exists on the stack-level, which has already been
associated with a port (serial, TCP, etc.)  As such, the only remaining
item is to define which virtual circuit ID will be used for the stack.
Then, of course, start the router.