Shifted testset to build using the included tinyxml sources.
Building with tinyxml as an external library was proving to be too
difficult.  The extremely 'tiny' size of tinyxml basically means
that trying to reuse a tinyxml shared object is an insane idea for
the effort it was taking.  So now testset will always be built and
deployed during a 'make install', which should be good/useful.