Expanded C++ demo code to allow parameters via CLI
When testing a custom DNP3 master, it can sometimes be beneficial to
start several slave instances with "known" behavior.  Since this quickly
becomes a chicken-and-egg problem, simply expanding the existing demo
apps to support changing basic parameters allows for the developer to
focus on their custom app rather than making both ends of the system.